Huge Update
Reworked how the file scanner and source files work. Source files no longer contain a dictionary with the line # label as the key. Instead, that's only handled in the str function. Also added and still working on the parser class and functions. Fleshed out but now need to run tests on all methods.